requirements.txt
fastapi
uvicorn
pytest
motor
pydantic
python-dotenv
pydantic[email]

